Improve the way to configure the test autoloading
Using autoload-dev rather than a bootstrap file to register the test
autoloading is easier.
The bootstrap file was also incompatible with composer-installed phpunit
because of its usage of _once to load the autoload.php file, which would
return true rather than the ClassLoader if the file was already loaded
previously.